Agrego validacion de nombre e identificador contra la DB al crear un nuevo pokemon y fix de query buscarPokemon

+++ CrearPokemon.php
@@ -3,27 +3,35 @@
 include_once ("Pokemon.php");
 include_once ("ConexionDatabase.php");
 
-//Muevo la imagen cargada a directorio img
-$rutaArchivoTemporal = $_FILES["imagen"]["tmp_name"];
-$rutaArchivoFinal = "img/" . $_FILES["imagen"]["name"];
-
-move_uploaded_file($rutaArchivoTemporal, $rutaArchivoFinal);
-
+$conexionDB = new ConexionDatabase();
 
 $imagen = 'img/' . $_FILES["imagen"]["name"];
 
 
-//Creo objeto pokemon
-$nuevoPokemon = new Pokemon($_POST["identificador"], $_POST["nombre"], $imagen, $_POST["tipo"], $_POST["descripcion"]);
 
-$conexionDB = new ConexionDatabase();
+//Verifico si existe el Pokemon
+$existePokemon = $conexionDB->existePokemon($_POST["identificador"], strtolower($_POST["nombre"]));
 
-$pokemonCreado = $conexionDB->crearPokemon($nuevoPokemon);
+if ($existePokemon) {
+    echo "<h2>Ha ocurrido un error!</h2><p>El nombre o número de Pokémon ya está en uso. Elija uno nuevo</p>".
+        "<a href='./vistas/VistaCrearPokemon.php'>Volver al formulario</a>";
+}
+else{
+    //Muevo la imagen cargada a directorio img
+    $rutaArchivoTemporal = $_FILES["imagen"]["tmp_name"];
+    $rutaArchivoFinal = "img/" . $_FILES["imagen"]["name"];
 
+    move_uploaded_file($rutaArchivoTemporal, $rutaArchivoFinal);
 
+    //Creo objeto pokemon
+    $nuevoPokemon = new Pokemon($_POST["identificador"], $_POST["nombre"], $imagen, $_POST["tipo"], $_POST["descripcion"]);
+    $pokemonCreado = $conexionDB->crearPokemon($nuevoPokemon);
 
-echo "<h2>Bien hecho! Creaste un nuevo Pokémon</h2>".
-     "<div> <h3>". $pokemonCreado->getNombre() . "</h3></div>
-       <img src=".$pokemonCreado->getimagen() ." width='20%'>".
+    echo "<h2>Bien hecho! Creaste un nuevo Pokémon</h2>".
+        "<div> <h3>". $pokemonCreado->getNombre() . "</h3></div>
+                 <img src=".$pokemonCreado->getimagen() ." width='20%'>".
         "<a href='index.php'>Volver a Home</a>";
+}
+
+
